Updated CMake project
CMake:
Updated the CMake project to work on latest Linuxes.
Dynamic creation of `.desktop` file on Linux.
Updated changelog file.
Removed dead code and old comments.

wxWidgets:
wxWidgets is searched for and if not found, downloaded and installed
automatically into the `CMAKE_EP_BASE_PATH`.

gettext:
The gettext dependency is downloaded and installed during CMake
configuration, if it is not found.
The project first searches for `Gettext` in the `PATH` environmental
variable.
Second, the `CMAKE_EP_BASE_PATH` is searched.
If not found, `Gettext` is automatically downloaded into the
`CMAKE_EP_BASE_PATH` folder.

Translations:
Translations are now found during development, as well as when
installed.

ReadMe:
Updated build instructions for Linux and Windows.
Details about dependencies and optional tools/arguments.